Include 'files' section of the server response
Often one needs information from the 'files' section of the Gist response. For example, it contains the raw URL of the created gist, which cannot be derived from other fields. Thus, it would be great to include this section in the output of the 'create' function, so that the user is able to employ this data straight away.